backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
openai
pandas
pandapipes
requests
langchain
langchain-openai
tiktoken
httpx
numpy
pydantic>=2
python-dotenv
networkx
matplotlib
orjson
pytest
starlette